feat: implement TaskList component (#4)
- Create TaskList component to manage task display
- Move task list logic from App.vue to dedicated component
- Implement empty state handling
- Use computed property for task list reactivity
- Maintain mobile-first design with Tailwind
